// src/frontend/src/src/backend.d.ts
export type Time = bigint;
export interface Attendance {
    status: AttendanceStatus;
    date: string;
    employeeId: bigint;
}
export interface DSAConnector {
    status: string;
    partnerName: string;
    name: string;
    email: string;
    phone: string;
}
export interface Transaction {
    id: bigint;
    transactionType: string;
    date: Time;
    description: string;
    category: string;
    amount: number;
}